drfToDoList/requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
drf_spectacular
django-cors-headers